+++ public/ide/modules/ide/js/Type.js
@@ -29,6 +29,8 @@ ideServices.
         var MyScriptTrigger=this.MyScriptTrigger='MyScriptTrigger';
         var MySlideBlock=this.MySlideBlock="MySlideBlock";
         var MyVideo=this.MyVideo="MyVideo";
+        var MyAnimation=this.MyAnimation="MyAnimation";
+
         var General = this.General = "general"
 
 
@@ -44,6 +46,7 @@ ideServices.
             _typeStr==this.MyProgress||_typeStr==this.MyNumber||_typeStr==this.MyButtonGroup||_typeStr == this.MyDashboard||
             _typeStr==this.MyKnob||_typeStr==this.MyTextArea||_typeStr==this.MyNum||_typeStr==this.MyOscilloscope||_typeStr==this.MyImage||
             _typeStr==this.MySwitch||_typeStr==this.MyRotateImg||_typeStr==this.MyDateTime||_typeStr==this.MyScriptTrigger||_typeStr==this.MySlideBlock||
+            _typeStr==this.MyVideo||_typeStr==this.MyAnimation){
             _typeStr==this.MyVideo||_typeStr==this.General){
             return true;
         }