+++ public/ide/modules/ide/js/parts/nav/dropdownController.js
@@ -15,7 +15,7 @@ ide.controller('addWidgetController', ['$scope', 'Type', function ($scope, Type)
         {name:"按钮组",icon:"&#xe637;"},
         {name:"触发器",icon:"&#xe634;"},
         {name:"影像",icon:"&#xe650;"},
-        {name:"G",icon:"&#xe650;"}
+        {name:"开机动画",icon:"&#xe633"}
     ];
     $scope.addWidget = function (index) {
         var widgets = $scope.widgets;