+++ public/ide/modules/ide/min-js/projectService.js
@@ -1039,22 +1039,18 @@ ideServices
 
                         syncSublayer(fabWidget);
                     },initiator);
-                }else if (_newWidget.type==Type.General){
-
-                    fabric.General.fromLevel(_newWidget, function (fabWidget) {
-                        _self.currentFabWidgetIdList=[fabWidget.id];
-
-
-
+                }else if(_newWidget.type==Type.MyAnimation){
+                    fabric.MyAnimation.fromLevel(_newWidget,function(fabWidget){
+                        _self.currentFabLayerIdList=[fabWidget.id];
+                        fabWidget.urls=_newWidget.subSlides;
                         subLayerNode.add(fabWidget);
                         subLayerNode.renderAll.bind(subLayerNode)();
+
                         _newWidget.info.width=fabWidget.getWidth();
                         _newWidget.info.height=fabWidget.getHeight();
 
                         syncSublayer(fabWidget);
                     },initiator);
-
-
                 }