+++ public/ide/modules/ide/min-js/widgetService.js
@@ -382,7 +382,6 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
 
                 self.backgroundColor=level.texList[0].slices[0].color;
                 self.backgroundImageElement=ResourceService.getResourceFromCache(level.texList[0].slices[0].imgSrc);
-                self.backgroundImageElement = ResourceService.getResourceFromCache(level.texList[0].slices[0].imgSrc);
                 if(self.progressModeId=='0'){
                     self.progressColor=level.texList[1].slices[0].color;
                     self.progressImageElement=ResourceService.getResourceFromCache(level.texList[1].slices[0].imgSrc);
@@ -484,8 +483,17 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
                 }else if(this.progressModeId=='1'){
                     //变色进度条
                     var progressColor = changeColor(this.initColor,this.endColor,this.progressValue);
-
                     //console.log(progressColor);
+                    ctx.fillStyle=this.backgroundColor;
+                    ctx.fillRect(
+                        -this.width / 2,
+                        -this.height / 2,
+                        this.width,
+                        this.height
+                    );
+                    if(this.backgroundImageElement){
+                        ctx.drawImage(this.backgroundImageElement, -this.width / 2, -this.height / 2,this.width,this.height);
+                    }
                     if(this.arrange=='horizontal'){
                         if(this.cursorImageElement){
                             ctx.drawImage(this.cursorImageElement,-this.width/2+(this.width*this.progressValue),-this.cursorImageElement.height/2/this.scaleY,this.cursorImageElement.width/this.scaleX,this.cursorImageElement.height/this.scaleY);
@@ -499,9 +507,6 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
                             ctx.drawImage(this.cursorImageElement,-this.cursorImageElement.width/2/this.scaleX,this.height/2-this.height*this.progressValue-this.cursorImageElement.height/this.scaleY,this.cursorImageElement.width/this.scaleX,this.cursorImageElement.height/this.scaleY);
                         }
                     }
-                    if(this.backgroundImageElement){
-                        ctx.drawImage(this.backgroundImageElement, -this.width / 2, -this.height / 2,this.width,this.height);
-                    }
                 }else if(this.progressModeId=='2'){
                     //脚本进度条，啥也不画！
                 }else if(this.progressModeId=='3'){
@@ -512,6 +517,9 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
                         this.width,
                         this.height
                     );
+                    if(this.backgroundImageElement){
+                        ctx.drawImage(this.backgroundImageElement, -this.width / 2, -this.height / 2,this.width,this.height);
+                    }
                     if(this.arrange=='horizontal'){
                         if(this.cursorImageElement){
                             ctx.drawImage(this.cursorImageElement,-this.width/2+(this.width*this.progressValue),-this.cursorImageElement.height/2/this.scaleY,this.cursorImageElement.width/this.scaleX,this.cursorImageElement.height/this.scaleY);
@@ -569,9 +577,6 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
                         if(this.cursorImageElement){
                             ctx.drawImage(this.cursorImageElement,-this.cursorImageElement.width/2/this.scaleX,this.height/2-this.height*this.progressValue-this.cursorImageElement.height/this.scaleY,this.cursorImageElement.width/this.scaleX,this.cursorImageElement.height/this.scaleY);
                         }
-                    }
-                    if(this.backgroundImageElement){
-                        ctx.drawImage(this.backgroundImageElement, -this.width / 2, -this.height / 2,this.width,this.height);
                     }
                 }
 
@@ -3073,90 +3078,76 @@ ideServices.service('WidgetService',['ProjectService', 'Type', 'ResourceService'
     };
     fabric.MySlide.async = true;
 
-    //general
-    fabric.General = fabric.util.createClass(fabric.Object, {
-        type: Type.General,
+    //myAnimation
+    fabric.MyAnimation = fabric.util.createClass(fabric.Object, {
+        type: Type.MyAnimation,
         initialize: function (level, options) {
             var self=this;
             this.callSuper('initialize',options);
             this.lockRotation=true;
             this.hasRotatingPoint=false;
-            // this.normalColor=level.texList[0].slices[0].color;
-            // this.arrange=level.info.arrange;
-            //
-            // this.text=level.info.text;
-            // this.fontFamily=level.info.fontFamily;
-            // this.fontSize=level.info.fontSize;
-            // this.fontColor=level.info.fontColor;
-            // this.fontBold=level.info.fontBold;
-            // this.fontItalic=level.info.fontItalic;
-            //
-            // this.normalImageElement = ResourceService.getResourceFromCache(level.texList[0].slices[0].imgSrc);
-            // if (this.normalImageElement) {
-            //     this.loaded = true;
-            //     this.setCoords();
-            //     this.fire('image:loaded');
-            // }
-            var x = level.info.left;
-            var y = level.info.top;
-            var w = level.info.width;
-            var h = level.info.height;
 
-            console.log(level.type)
-            this.curWidget = new WidgetModel.models['Button'](x,y,w,h,'button',null,level.texList[0].slices)
-            this.curWidgetInfo = this.curWidget.toObject()
-            // console.log(this.curWidgetInfo)
-            // console.log(WidgetModel.WidgetCommandParser.complier.transformer.trans(WidgetModel.WidgetCommandParser.complier.parser.parse(this.curWidgetInfo.onInitialize)))
-            // var functionBody = WidgetModel.WidgetCommandParser.transFunction(this.curWidgetInfo,this.curWidgetInfo.onInitialize);
-            // console.log(functionBody)
-            // this.curWidgetInfo.onInitialize = new Function(functionBody)
-            // this.curWidgetInfo.onInitialize();
+            var tex=level.texList[0];
+            this.currentColor=tex.slices[tex.currentSliceIdx].color;
+
+            this.currentImageElement = ResourceService.getResourceFromCache(tex.slices[tex.currentSliceIdx].imgSrc);
+            if (this.currentImageElement) {
+                this.loaded = true;
+                this.setCoords();
+                this.fire('image:loaded');
+            }
+            this.on('changeTex', function (arg) {
+
+                var level=arg.level;
+                var _callback=arg.callback;
+
+                var tex=level.texList[0];
+                self.currentColor=tex.slices[tex.currentSliceIdx].color;
+                self.currentImageElement = ResourceService.getResourceFromCache(tex.slices[tex.currentSliceIdx].imgSrc);
+                var subLayerNode=CanvasService.getSubLayerNode();
+                subLayerNode.renderAll();
+                _callback&&_callback();
+            });
+
         },
         toObject: function () {
             return fabric.util.object.extend(this.callSuper('toObject'));
         },
         _render: function (ctx) {
             try{
-
-                // console.log('drawing general',this.level,this)
-                var firstLayer = this.curWidgetInfo.layers[0];
-                ctx.fillStyle=firstLayer.subLayers.color.color;
-                var info = this.curWidgetInfo.info;
+                ctx.fillStyle=this.currentColor;
                 ctx.fillRect(
-                    -(info.left / 2),
-                    -(info.top / 2) ,
-                    info.width,
-                    info.height );
-                // if (this.currentImageElement){
-                //     ctx.drawImage(this.currentImageElement, -this.width / 2, -this.height / 2,this.width,this.height);
-                //
-                // }
+                    -(this.width / 2),
+                    -(this.height / 2) ,
+                    this.width ,
+                    this.height );
+                if (this.currentImageElement){
+                    ctx.drawImage(this.currentImageElement, -this.width / 2, -this.height / 2,this.width,this.height);
+
+                }
             }
             catch(err){
                 console.log('错误描述',err);
-                toastr.warning('渲染按钮出错');
+                toastr.warning('渲染Slide出错');
             }
         }
     });
-    fabric.General.fromLevel= function (level, callback,option) {
-        callback && callback(new fabric.General(level, option));
-    };
-    fabric.General.prototype.toObject = (function (toObject) {
+    fabric.MyAnimation.fromLevel= function (level, callback,option) {
+        callback && callback(new fabric.MyAnimation(level, option));
+    }
+    fabric.MyAnimation.prototype.toObject = (function (toObject) {
         return function () {
             return fabric.util.object.extend(toObject.call(this), {
-                // normalImageElement:this.normalImageElement,
-                // normalColor:this.normalColor
+                currentColor:this.currentColor,
+                currentImageElement:this.currentImageElement
             });
         }
-    })(fabric.General.prototype.toObject);
-    fabric.General.fromObject = function (object, callback) {
+    })(fabric.MyAnimation.prototype.toObject);
+    fabric.MyAnimation.fromObject = function (object, callback) {
         var level=ProjectService.getLevelById(object.id);
-        callback && callback(new fabric.General(level, object));
+        callback && callback(new fabric.MyAnimation(level, object));
     };
-    fabric.General.async = true;
+    fabric.MyAnimation.async = true;
 
 
-
-
-
-}])
+}]);